// app/src/app/enums/error-codes.ts
// there should be proper error message for each of these in lang files
export enum ERROR_CODES {
  NOT_FOUND = 'NF404',
  NOT_AUTHENTICATED = 'NAUTH',
  UNEXPECTED_ERROR = 'NTEXP',
  NOT_AUTHORIZED = 'NAUZE',
  PROJECT_ALREADY_FAVORITED = 'PRFAV',
  PROJECT_ALREADY_VOTED= "PRVOT",
  ACCOUNT_ALREADY_EXISTS = 'NEXIS',
  INVALID_CREDENTIALS = 'NCRED',
  IMAGE_UPLOAD = 'IMGUP',
  IMAGE_REMOVE = 'IMGRM',
  PROJECT_NOT_FOUND = 'PRJNF',
  BAD_REQUEST = 'BDREQ',
  BANNED = 'BANND',
  INVALID_GROUP= 'NOGRP',
}